requirements.txt
Flask
openai
gunicorn
python-decouple
spacy